use build in node function if possible add async also on generator e2e split also e2e for init in separate file.
