+++ test/generate-odata-client.e2e.spec.ts
@@ -1,6 +1,8 @@
 /*!
  * Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved.
  */
+import { deleteAsync } from './test-utils';
+
 jest.mock('cli-ux', () => ({
   default: {
     confirm: jest.fn().mockResolvedValue(true)
@@ -10,21 +12,20 @@ jest.retryTimes(3);
 
 import * as fs from 'fs-extra';
 import * as path from 'path';
-import * as rm from 'rimraf';
 import GenerateODataClient from '../src/commands/generate-odata-client';
 
 describe('generate-odata-client', () => {
   const pathForTests = path.resolve(__dirname, __filename.replace(/\./g, '-')).replace('-ts', '');
 
-  beforeAll(() => {
-    rm.sync(pathForTests);
+  beforeAll(async () => {
+    await deleteAsync(pathForTests, 6);
     const pathForResources = path.resolve(__dirname, 'resources', 'template-generator-odata-client');
     fs.copySync(pathForResources, pathForTests);
-  });
+  }, 80000);
 
-  afterAll(() => {
-    rm.sync(pathForTests);
-  });
+  afterAll(async () => {
+    return deleteAsync(pathForTests, 6);
+  }, 80000);
 
   test('[E2E] should generate a OData client', async () => {
     const result = await GenerateODataClient.run(['-i', 'edmxSource', '-o', 'output', '--projectDir', pathForTests]);
